+++ src/extension.ts
@@ -132,13 +132,7 @@ export  function activate(context: vscode.ExtensionContext) {
 		const componentName = await vscode.window.showInputBox({ placeHolder: 'Enter component name' });
 		const newComponentPath = vscode.workspace.rootPath + '/src/components/' + componentName + ext;
 		fs.copyFileSync(__dirname + '/stubs/' + framework?.toLowerCase() + '/' + filename, newComponentPath);
-		const editor = vscode.window.activeTextEditor;
-		const selection = editor?.selection;
-		if (selection && !selection.isEmpty) {
-			const selectionRange = new vscode.Range(selection.start.line, selection.start.character, selection.end.line, selection.end.character);
-			const highlighted = editor.document.getText(selectionRange);
-			console.log(highlighted);
-		}
+		
 		if(componentName){
 			fs.writeFileSync(newComponentPath, fs.readFileSync(newComponentPath, 'utf8').replace(/ComponentName/g, componentName), 'utf8')
 		}
@@ -147,8 +141,69 @@ export  function activate(context: vscode.ExtensionContext) {
 
 			
 	})
-	context.subscriptions.push(disposable, quickCommand);
+
+
+
+	let componentFromSelectionCommand = vscode.commands.registerCommand('make-js-component-vscode.make-js-component-from-selection', async () => {	
+		let filename = '';
+		let ext = '.js';
+		let framework = detectFramework();
+		let useTs = false;
+		if(["React", "Angular","Svelte"].includes(framework||'')){
+			useTs = await vscode.window.showQuickPick(["Yes", "No"],{placeHolder: "Use Typescript?"}) === "Yes";
+		}
+		switch (framework) {
+			case 'Vue':
+				let type = await vscode.window.showQuickPick(["Options API", "Composition API"], { placeHolder: 'Select type' });
+				type = type?.split(' ')[0].toLocaleLowerCase();
+				filename = `component-${type}.vue.stub`;
+				ext = '.vue';
+				break;
+			case 'React':
+				ext = useTs ? '.tsx' : '.jsx';
+				filename = `function-component${ext}.stub`;
+				break;
+			case 'Astro':
+				filename = 'component.astro.stub';
+				ext = '.astro';
+				break;
+			case 'Svelte':
+				const ts = useTs ? 'ts' : 'js';
+				filename = `component-${ts}.svelte.stub`;
+				ext = '.svelte';
+				break;
+			case 'Qwik':
+				filename = 'hello-world-component.tsx.stub';
+				ext = '.qwik.ts';
+				break;
+			case 'Angular':
+				ext = useTs ? '.ts' : '.js';
+				filename = `component.component${ext}.stub`;
+				break;
+			default:
+				break;
+		}
+		const componentName = await vscode.window.showInputBox({ placeHolder: 'Enter component name' });
+		const newComponentPath = vscode.workspace.rootPath + '/src/components/' + componentName + ext;
+		fs.copyFileSync(__dirname + '/stubs/' + framework?.toLowerCase() + '/' + filename, newComponentPath);
+		const editor = vscode.window.activeTextEditor;
+		const selection = editor?.selection;
+		if (selection && !selection.isEmpty) {
+			const selectionRange = new vscode.Range(selection.start.line, selection.start.character, selection.end.line, selection.end.character);
+			const highlighted = editor.document.getText(selectionRange);
+			if(componentName){
+				fs.writeFileSync(newComponentPath, fs.readFileSync(newComponentPath, 'utf8').replace(/Hello ComponentName/g, highlighted), 'utf8')
+			}
+		}
+		vscode.workspace.openTextDocument(newComponentPath).then(doc => vscode.window.showTextDocument(doc));
+		vscode.window.showInformationMessage(`Component ${componentName} created!`);
+
+			
+	})
+
+	context.subscriptions.push(disposable, quickCommand,componentFromSelectionCommand);
 }
+