+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "make-js-component-vscode" extension will be document
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 ## [unreleased]
+- create component frokm selection
 ## [0.0.5]
 - pick between Js and Ts for frameworks that need it
 - pick between composition and option api for vue